backend/live-bridge/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
google-adk
websockets>=13.0
